requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0

sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.3.0

loguru==0.7.2
httpx>=0.27.0,<0.29.0

argon2-cffi>=23.1.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-telegram-bot>=20.7